refactor(back-end): add return before each response and use if else to check each condition

+++ back-end/controllers/record-controller.js
@@ -27,7 +27,7 @@ const recordController = {
         },
       })
 
-      res.status(201).end()
+      return res.status(201).end()
       // #swagger.tags = ['Expense Record']
     } catch (error) {
       next(error)
@@ -41,6 +41,7 @@ const recordController = {
       const recordId = Number(req.params.rid)
       const { title, amount, note, categoryId } = req.body
       const userId = req.user.id
+      console.log(userId)
 
       // Check if there are missing data
       if (!title || !amount || !categoryId)
@@ -54,14 +55,13 @@ const recordController = {
           id: recordId,
         },
       })
-      theRecord ? theRecord : res.status(400).json('the record does not exist')
+      if (!theRecord) return res.status(400).json('the record does not exist')
 
       // Check if the record is one of the current user's records
-      userId === theRecord.userId
-        ? userId
-        : res
-            .status(400)
-            .json('this record does not belong to the current user')
+      if (userId !== theRecord.userId)
+        return res
+          .status(400)
+          .json('this record does not belong to the current user')
 
       // Update this record
       await prisma.record.update({
@@ -69,7 +69,7 @@ const recordController = {
         data: { title, amount, note, categoryId },
       })
 
-      res.status(204).json('successfully update this record')
+      return res.status(204)
       // #swagger.tags = ['Expense Record']
     } catch (error) {
       next(error)
@@ -89,14 +89,13 @@ const recordController = {
           id: recordId,
         },
       })
-      theRecord ? theRecord : res.status(400).json('the record does not exist')
+      if (!theRecord) return res.status(400).json('the record does not exist')
 
       // Check if the record is one of the current user's records
-      userId === theRecord.userId
-        ? userId
-        : res
-            .status(400)
-            .json('this record does not belong to the current user')
+      if (userId !== theRecord.userId)
+        return res
+          .status(400)
+          .json('this record does not belong to the current user')
 
       // Delete this record
       await prisma.record.delete({
@@ -105,7 +104,7 @@ const recordController = {
         },
       })
 
-      res.status(204).json('successfully delete this record')
+      return res.status(204)
       // #swagger.tags = ['Expense Record']
     } catch (error) {
       next(error)
@@ -138,7 +137,7 @@ const recordController = {
           categoryIds
         )});`
 
-      res.status(200).json(expenseRecords)
+      return res.status(200).json(expenseRecords)
       // #swagger.tags = ['Expense Record']
     } catch (error) {
       next(error)
@@ -171,7 +170,7 @@ const recordController = {
           categoryIds
         )});`
 
-      res.status(200).json(incomeRecords)
+      return res.status(200).json(incomeRecords)
     } catch (error) {
       next(error)
     }
@@ -197,7 +196,7 @@ const recordController = {
           categoryIds
         )});`
 
-      res.status(200).json(savingRecords)
+      return res.status(200).json(savingRecords)
     } catch (error) {
       next(error)
     }
